+++ src/renderers/atomics/Embed/index.jsx
@@ -1,34 +1,20 @@
 import './style.scss'
 import React from 'react'
-import StaticContainer from 'components/common/StaticContainer'
+import PlayerModal from 'components/business/PlayerModal'
 import { ContentUtils } from 'braft-utils'
 
 export default class Embed extends React.Component {
 
-  state = {
-    toolbarVisible: false,
-  }
-
   render () {
 
-    const { toolbarVisible } = this.state
-    const { mediaData } = this.props
-    const { url } = mediaData
+    const { mediaData, language } = this.props
+    const { name, url, meta } = mediaData
 
     return (
-      <div
-        className='bf-embed'
-        onMouseOver={this.showToolbar}
-        onMouseLeave={this.hideToolbar}
-      >
-        <StaticContainer>
+      <div className='bf-embed-wrap'>
+        <PlayerModal type="embed" onRemove={this.removeEmbed} poster={meta ? meta.poster || '' : ''} language={language} url={url} name={name} title={language.videoPlayer.embedTitle}>
           <div className='bf-embed-player' dangerouslySetInnerHTML={{ __html: url}}></div>
-        </StaticContainer>
-        {toolbarVisible ? (
-          <div className='bf-media-toolbar'>
-            <a onClick={this.removeEmbed}>&#xe9ac;</a>
-          </div>
-        ) : null}
+        </PlayerModal>
       </div>
     )
 
@@ -36,18 +22,6 @@ export default class Embed extends React.Component {
 
   removeEmbed = () => {
     this.props.editor.setValue(ContentUtils.removeBlock(this.props.editorState, this.props.block))
-  }
-
-  showToolbar = () => {
-    this.setState({
-      toolbarVisible: true
-    })
-  }
-
-  hideToolbar = () => {
-    this.setState({
-      toolbarVisible: false
-    })
   }
 
 }